+++ CHANGELOG.md
@@ -74,7 +74,7 @@ All notable changes to this project will be documented in this file.
   - `toPositiveIntOrNull` and `toPositiveIntOrThrow`
   - `toNegativeIntOrNull` and `toNegativeIntOrThrow`
   - `toStrictlyPositiveIntOrNull` and `toStrictlyPositiveIntOrThrow`
-  - `toStrictlyNegativeIntOrNull`.
+  - `toStrictlyNegativeIntOrNull` and `toStrictlyNegativeIntOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258